+++ openghg/analyse/__init__.py
@@ -1,10 +1,10 @@
 from ._alignment import combine_datasets
 from ._fp_x_flux import (
     align_flux_to_time_targets,
-    fp_x_flux_keep_space,
-    fp_x_flux_keep_space_core,
+    fp_x_flux,
+    fp_x_flux_core,
     warm_numba_fp_x_flux,
-    write_fp_x_flux_keep_space_zarr,
+    write_fp_x_flux_zarr,
 )
 from ._modelled_obs import fp_x_flux_integrated, make_integrated_low_freq_flux
 from ._scenario import ModelScenario